initial placement changes for reset, search, filters - not styled
