+++ civictechprojects/static/css/partials/_ProjectFilterContainer.scss
@@ -25,9 +25,6 @@
   display: flex;
   justify-content: space-between;
 }
-.ProjectFilterContainer-root > div {
-  display: block;
-}
 
 .ProjectFilterContainer-filter-list {
   list-style-type: none;
@@ -95,4 +92,10 @@
 .ProjectFilterContainer-showtext {
   font-size: 16px;
   font-weight: normal;
+}
+
+.ProjectFilterContainer-reset {
+  width: 100%;
+  display: flex;
+  justify-content: center;
 }
