+++ common/components/componentsBySection/FindProjects/Filters/RenderFilterCategory.jsx
@@ -15,7 +15,6 @@ const classCategoryExpanded = 'ProjectFilterContainer-category ProjectFilterCont
 const classCategoryCollapsed = 'ProjectFilterContainer-category ProjectFilterContainer-collapsed';
 const classSubcategoryExpanded = 'ProjectFilterContainer-subcategory ProjectFilterContainer-expanded';
 const classSubcategoryCollapsed = 'ProjectFilterContainer-subcategory ProjectFilterContainer-collapsed';
-const tagKeyPattern = tag.category + '-' + tag.display_name
 
 class RenderFilterCategory<T> extends React.PureComponent<Props<T>, State> {
   constructor(props: Props): void {
@@ -105,9 +104,9 @@ class RenderFilterCategory<T> extends React.PureComponent<Props<T>, State> {
   _renderFilterList(data) {
     //this function renders individual clickable filter items regardless of category or subcategory status
     let sortedTags = Object.values(data).map((tag) =>
-      <li key={tagKeyPattern} className="ProjectFilterContainer-list-item">
-          <input type="checkbox" id={tagKeyPattern} checked={this.props.checkEnabled(tag)} onChange={() => this.props.selectOption(tag)}></input>
-          <label htmlFor={tagKeyPattern}>
+      <li key={tag.category + '-' + tag.display_name} className="ProjectFilterContainer-list-item">
+          <input type="checkbox" id={tag.category + '-' + tag.display_name} checked={this.props.checkEnabled(tag)} onChange={() => this.props.selectOption(tag)}></input>
+          <label htmlFor={tag.category + '-' + tag.display_name}>
             <span className="ProjectFilterContainer-list-item-name">{tag.display_name}</span> <span className="ProjectFilterContainer-list-item-count">{this.props.checkEnabled(tag) ? <i className={GlyphStyles.Check}></i> : tag.num_times}</span>
           </label>
       </li>
